frontend/narration_requirements.txt
# ElevenLabs Narration Generator

elevenlabs==0.2.27
python-dotenv==1.0.0
